feat(runtime): end-to-end skill dispatch integration test (Plan 4 Task 11)
Add `crates/athena-voice-runtime/tests/skill_dispatch.rs` driving the
smoke-test WASM through the router + SkillDispatcher + TTS actor DAG
and asserting `Event::IntentMatched`, `Event::SkillInvoked`, TTS chunks
carrying the skill's speech, and the absence of `Event::LlmFallback`.

Fixes to make the test green:

- Router: accept an optional `SkillDispatcherHandle` + a direct
  `tts_tok_tx` bypass; on match with a dispatcher wired, dispatch the
  skill inline, forward `SkillResponse::Speak` text to TTS (with a
  trailing sentence boundary so the buffered TTS actor flushes), and
  skip the LLM fallback path. Without a dispatcher, keep the legacy
  behaviour of emitting `SkillInvoked` + `LlmFallback`.
- SatelliteDeps: plumb the optional dispatcher through to `RouterDeps`.
- `host_mqtt_publish` ABI: `extism-pdk`'s guest-side `host_fn!` widens
  every non-void return to a `u64` memory handle and decodes bytes via
  `FromBytesOwned`, so a scalar `Val::I64(code)` return was mis-parsed
  as a memory offset ("could not convert slice to array"). Return the
  little-endian `i64` bytes through `plugin.memory_new(...)` instead,
  matching the pattern used by `host_config_get` / `host_state_get`.
  Widen the guest-side `-> i32` return + constants to `i64` so the ABI
  matches on both sides.
- Derive `Clone` on `RuleIndex` so the test can share the registry's
  pattern index with the router while retaining ownership of the
  registry inside the dispatcher.

`cargo test --workspace` green (139 tests); `cargo clippy --workspace
--all-targets -- -D warnings` clean.

+++ crates/athena-voice-skill-sdk/src/host.rs
@@ -75,9 +75,9 @@ mod guest {
 
     // Result codes mirror `MQTT_OK` / `MQTT_ERR_ACL` / `MQTT_ERR_CLIENT` in
     // `crates/athena-voice-runtime/src/wasm/host_fns.rs`.
-    const MQTT_OK: i32 = 0;
-    const MQTT_ERR_ACL: i32 = 1;
-    const MQTT_ERR_CLIENT: i32 = 2;
+    const MQTT_OK: i64 = 0;
+    const MQTT_ERR_ACL: i64 = 1;
+    const MQTT_ERR_CLIENT: i64 = 2;
 
     #[host_fn]
     unsafe extern "ExtismHost" {
@@ -85,7 +85,7 @@ mod guest {
         fn host_config_get(key: String) -> String;
         fn host_state_get(key: String) -> Vec<u8>;
         fn host_state_set(key: String, val: Vec<u8>);
-        fn host_mqtt_publish(topic: String, payload: Vec<u8>) -> i32;
+        fn host_mqtt_publish(topic: String, payload: Vec<u8>) -> i64;
         fn host_http_get_json(url: String) -> Vec<u8>;
     }
